Reorganized file structure updated reccomended items
