--- src/api/account.types.ts
@@ -1,1 +1,0 @@
-export type AccountId = { accId: number };
