+++ src/api/accounts/index.ts
@@ -1,0 +1,6 @@
+import getAccountInfo from './getInformation';
+import updateStatus from './updateStatus';
+import deleteSession from './deleteSession';
+
+
+export  { getAccountInfo, updateStatus, deleteSession };
