+++ src/index.ts
@@ -1,36 +1,26 @@
 import { initDbConnection, closeDbConnection } from './db';
-import {
-  getOnlineCharacters,
-  getUniquePlayerCount,
-  getIdsByCharName,
-  unstuckCharacter,
-} from './api/characters';
-
-import {
-  getAccountIdByUserName,
-  banPlayerByAccId,
-  unbanPlayerByAccId,
-  killPlayerSession,
-} from './api/account';
-
-import { getTotalServerGil, getTopTenAccountsByGil, getTopTenCharactersByGil } from './api/audit';
+import * as accounts from './api/accounts';
+import * as audits from './api/audits';
+import * as characters from './api/characters';
 
 const init = async () => {
   initDbConnection();
 
-  // console.log(await getOnlineCharacters());
-  // console.log(await getUniquePlayerCount());
-  // console.log(await getIdsByCharName('Flam'));
-  // console.log(await getIdsByCharName('Flammm'));
-  // console.log(await unstuckCharacter('Flam'));
-  // console.log(await unstuckCharacter('Flffam'));
-  // console.log(await getAccountIdByUserName('Bob'))
-  // console.log(await banPlayerByAccId('Bob'))
-  // console.log(await unbanPlayerByAccId('Bob'));
-  // console.log(await killPlayerSession('Bob'));
-  // console.log(await getTotalServerGil());
-  console.log(await getTopTenAccountsByGil());
-  console.log(await getTopTenCharactersByGil());
+  // Accounts Tests
+  // console.log(await accounts.getAccountInfo('Bob'));
+  // console.log(await accounts.getAccountInfo(1004));
+  // console.log(await accounts.updateStatus('Testy', 'Active'));
+  // console.log(await accounts.deleteSession('Test'));
+
+  // Audit Tests
+  // console.log(await audits.getTotalServerGil());
+  // console.log(await audits.getTopGilByAccount(5));
+  // console.log(await audits.getTopGilByCharacter(2));
+
+  // Character Tests
+  // console.log(await characters.getOnlineCharacters(1));
+  // console.log(await characters.getUniqueCharacterCount());
+  // console.log(await characters.unstuckCharacter('Atest'));
 
   closeDbConnection(() => {
     console.log('Closed the database connection');
